Solved!

+++ assignments/08Caracol/src/exercise.py
@@ -1,7 +1,7 @@
 def main():
     #escribe tu código abajo de esta línea
-    minutos= int(input("Dame los minutos: "))
-    centimetros= ((60*5.7) / 10) * minutos
+    minutos= float(input("Dame los minutos: "))
+    centimetros= (((60*5.7) / 10) * minutos)
     
     print("Centímetros recorridos:", centimetros)
